server/src/requirements.txt
Flask==2.1.2
Werkzeug==2.2.2
python-dotenv
aws-wsgi
pynacl
arrow

